moved things around so they work under version control
