+++ robots/Tank.java
@@ -1,15 +1,15 @@
-package nsl.robots;
+package nslbattlecodemaster.robots;
 
 import battlecode.common.*;
 
 /**
  * Created by morris on 1/11/17.
  */
-public strictfp class Soldier extends NSLRobot {
+public strictfp class Tank extends NSLRobot {
 
-    public Soldier(RobotController rc) {
+    public Tank(RobotController rc) {
         super(rc);
-        name = "Soldier";
+        name = "Tank";
     }
 
     @SuppressWarnings("unused")
